+++ experiments/random/random_experiment.py
@@ -1,4 +1,4 @@
-from code.algorithms import randomise
+from radio_russia.algorithms import randomise
 import matplotlib.pyplot as plt
 import copy
 import csv
