+++ radio_russia/algorithms/greedy.py
